+++ src/composables/navigraph.ts
@@ -137,6 +137,7 @@ export async function getNavigraphAirportProcedure<T extends NavigraphDataAirpor
 
 const replacementRegex = /[^a-zA-Z0-9\/]+/;
 const latRegex = /^(\d{2,4})([NS])/;
+const lonRegex = /^(\d{3,5})([EW])/;
 const sidstarRegex = /(?<start>[A-Z]{4})([A-Z]?)(?<end>[0-9][A-Z])/;
 
 export interface EnroutePath {
@@ -162,7 +163,7 @@ function getPreciseCoord(input: string): [Coordinate, string] | null {
 
     const remainder = input.slice(latMatch[0].length);
 
-    const lonMatch = remainder.match(/^(\d{3,5})([EW])/);
+    const lonMatch = remainder.match(lonRegex);
     if (!lonMatch) return null;
 
     const lonDigits = lonMatch[1];
@@ -211,6 +212,8 @@ export async function getFlightPlanWaypoints({ flightPlan, departure, arrival }:
         for (let i = 0; i < entries.length; i++) {
             const entry = entries[i];
             let split = entry.split('/');
+            if (split.length > 2) split = split.slice(split.length - 2, split.length);
+            if (split[1] && latRegex.test(split[1])) split.reverse();
             const search = split[0];
 
             if (split[1] && entry.startsWith(departure)) depRunway = split[1];
@@ -225,9 +228,6 @@ export async function getFlightPlanWaypoints({ flightPlan, departure, arrival }:
                 const sid = sids.findIndex(x => x.identifier === `${ tested?.groups?.start }${ tested?.groups?.end }`);
 
                 if (sid !== -1) {
-                    // We already have it fully drawn
-                    if (dataStore.navigraphProcedures[departure]?.sids[sids[sid].identifier]) continue;
-
                     const procedure = await getNavigraphAirportProcedure('sids', departure, sid);
 
                     if (depRunway) {
@@ -238,6 +238,12 @@ export async function getFlightPlanWaypoints({ flightPlan, departure, arrival }:
                                 title: procedure?.procedure.identifier,
                                 coordinate: x.coordinate,
                                 kind: 'sids',
+
+                                altitude: x.altitude,
+                                altitude1: x.altitude1,
+                                altitude2: x.altitude2,
+                                speed: x.speed,
+                                speedLimit: x.speedLimit,
                             } satisfies NavigraphNavDataEnrouteWaypointPartial)));
                         }
                     }
@@ -247,6 +253,12 @@ export async function getFlightPlanWaypoints({ flightPlan, departure, arrival }:
                         identifier: x.identifier,
                         coordinate: x.coordinate,
                         kind: 'sids',
+
+                        altitude: x.altitude,
+                        altitude1: x.altitude1,
+                        altitude2: x.altitude2,
+                        speed: x.speed,
+                        speedLimit: x.speedLimit,
                     } satisfies NavigraphNavDataEnrouteWaypointPartial)) ?? []);
 
                     const enrouteTransition = procedure?.transitions.enroute.find(x => x.name === entries[1] || x.name === entries[2]);
@@ -256,6 +268,12 @@ export async function getFlightPlanWaypoints({ flightPlan, departure, arrival }:
                             identifier: x.identifier,
                             coordinate: x.coordinate,
                             kind: 'sids',
+
+                            altitude: x.altitude,
+                            altitude1: x.altitude1,
+                            altitude2: x.altitude2,
+                            speed: x.speed,
+                            speedLimit: x.speedLimit,
                         } satisfies NavigraphNavDataEnrouteWaypointPartial)));
                     }
 
@@ -279,9 +297,6 @@ export async function getFlightPlanWaypoints({ flightPlan, departure, arrival }:
                 if (nextEntryTest && stars.some(x => x.identifier === `${ nextEntryTest.groups?.start }${ nextEntryTest.groups?.end }`)) continue;
 
                 if (star !== -1) {
-                    // We already have it fully drawn
-                    if (dataStore.navigraphProcedures[arrival]?.stars[stars[star].identifier]) continue;
-
                     starInit = true;
 
                     const procedure = await getNavigraphAirportProcedure('stars', arrival, star);
@@ -294,6 +309,12 @@ export async function getFlightPlanWaypoints({ flightPlan, departure, arrival }:
                             identifier: x.identifier,
                             coordinate: x.coordinate,
                             kind: 'stars',
+
+                            altitude: x.altitude,
+                            altitude1: x.altitude1,
+                            altitude2: x.altitude2,
+                            speed: x.speed,
+                            speedLimit: x.speedLimit,
                         } satisfies NavigraphNavDataEnrouteWaypointPartial)));
                     }
 
@@ -302,6 +323,12 @@ export async function getFlightPlanWaypoints({ flightPlan, departure, arrival }:
                         identifier: x.identifier,
                         coordinate: x.coordinate,
                         kind: 'stars',
+
+                        altitude: x.altitude,
+                        altitude1: x.altitude1,
+                        altitude2: x.altitude2,
+                        speed: x.speed,
+                        speedLimit: x.speedLimit,
                     } satisfies NavigraphNavDataEnrouteWaypointPartial)) ?? []);
 
                     if (arrRunway) {
@@ -312,6 +339,12 @@ export async function getFlightPlanWaypoints({ flightPlan, departure, arrival }:
                                 identifier: x.identifier,
                                 coordinate: x.coordinate,
                                 kind: 'stars',
+
+                                altitude: x.altitude,
+                                altitude1: x.altitude1,
+                                altitude2: x.altitude2,
+                                speed: x.speed,
+                                speedLimit: x.speedLimit,
                             } satisfies NavigraphNavDataEnrouteWaypointPartial)));
                         }
                     }
@@ -326,6 +359,12 @@ export async function getFlightPlanWaypoints({ flightPlan, departure, arrival }:
                                     identifier: x.identifier,
                                     coordinate: x.coordinate,
                                     kind: 'approaches',
+
+                                    altitude: x.altitude,
+                                    altitude1: x.altitude1,
+                                    altitude2: x.altitude2,
+                                    speed: x.speed,
+                                    speedLimit: x.speedLimit,
                                 } satisfies NavigraphNavDataEnrouteWaypointPartial)));
                             }
 
@@ -333,6 +372,12 @@ export async function getFlightPlanWaypoints({ flightPlan, departure, arrival }:
                                 identifier: x.identifier,
                                 coordinate: x.coordinate,
                                 kind: 'approaches',
+
+                                altitude: x.altitude,
+                                altitude1: x.altitude1,
+                                altitude2: x.altitude2,
+                                speed: x.speed,
+                                speedLimit: x.speedLimit,
                             } satisfies NavigraphNavDataEnrouteWaypointPartial)));
 
                             if (procedure.procedure.missedApproach) {
@@ -340,6 +385,12 @@ export async function getFlightPlanWaypoints({ flightPlan, departure, arrival }:
                                     identifier: x.identifier,
                                     coordinate: x.coordinate,
                                     kind: 'missedApproach',
+
+                                    altitude: x.altitude,
+                                    altitude1: x.altitude1,
+                                    altitude2: x.altitude2,
+                                    speed: x.speed,
+                                    speedLimit: x.speedLimit,
                                 } satisfies NavigraphNavDataEnrouteWaypointPartial)));
                             }
                         }