+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+# [1.2.0-beta.3]
+
+## QA
+
+- Revered random reloads fix
+- Fixed full route shown for some reason when you enable star for aircraft
+- Fixed duplicate waypoints display
+- Added constraints to predicted star/sid/approach
+- Removed line to airport from predicted missed approach
+- Improved aircraft predicted route update speed
+- Added holdings to enroute waypoints
+- Implemented route for all airport aircraft
+
+## Public Release
+
+- New icons from DotWallop: B105, BE20, BR23, BT7, C2, C25B, C414, E2, F4, F104, G91, H60, H64, HUNT, IL96, LANC, ME09, ME62, MIR2, S92, SB39, STAR
+- Improved icon for FA8X
+- Added model matching for C207, C30J and VF35
+
 # [1.2.0-beta.2]
 
 ## QA
@@ -8,9 +27,8 @@
 
 ## Public release
 
-- Removed random VATSIM Radar restarts when you open it
-- Fixed random website reloads when you open it
 - Added new precipitation weather layer - provided by OpenWeather. Old has been renamed to "Precipitation intensity"
+
 
 # [1.2.0-beta.1]